+++ scripts/install-deps.sh
@@ -654,6 +654,7 @@ pkgs_bench+=( iperf )
 pkgs_bench+=( fio )
 
 pkgs_media+=( vlc )
+pkgs_media+=( translate-shell )
 # https://motrix.app/download
 # https://code.visualstudio.com/docs/setup/linux
 # https://www.drawio.com/, https://github.com/jgraph/drawio-desktop/releases/